requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
scikit-learn>=1.3
mlflow>=2.10
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
fastapi>=0.100
uvicorn>=0.20
pydantic>=2.0
shap>=0.45
joblib>=1.3
apache-airflow>=2.8
great_expectations>=1.0
